+++ plugins/gtm-base/lib/gtmbase/__init__.py
@@ -1,3 +1,3 @@
 """GTM Base shared library."""
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"